requirements.txt
streamlit==1.38.0
ultralytics==8.3.32
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
torch==2.5.1